requirements.txt
kornia=0.6.12=pypi_0
numpy=1.21.2=py37h20f2e39_0
numpy-base=1.21.2=py37h79a1101_0
python=3.7.11=h12debd9_0
pytorch=1.10.0=py3.7_cuda11.3_cudnn8.2.0_0
thop=0.1.1-2209072238=pypi_0
torchaudio=0.10.0=py37_cu113
torchvision=0.11.1=py37_cu113


